Move transport from physics to its own subpackage
This commit includes changes made using PyCharm to automatically fix
references.  The documentation still needs to be updated.
