+++ plasmapy/physics/__init__.py
@@ -36,6 +36,6 @@ from .quantum import (deBroglie_wavelength,
 
 from .relativity import Lorentz_factor
 
-from . import transport
+from plasmapy import transport
 
 from . import magnetostatics
