+++ plasmapy/transport/tests/test_transport.py
@@ -12,36 +12,36 @@ from plasmapy.utils.exceptions import (PhysicsError,
 from plasmapy.physics.parameters import Hall_parameter
 from plasmapy.constants import m_p, m_e
 
-from plasmapy.physics.transport.collisions import (Coulomb_logarithm,
+from plasmapy.transport import (Coulomb_logarithm,
                                                    )
-from plasmapy.physics.transport.braginskii import (_nondim_thermal_conductivity,
-                                                   _nondim_viscosity,
-                                                   _nondim_resistivity,
-                                                   _nondim_te_conductivity,
-                                                   _check_Z,
-                                                   _nondim_tc_e_spitzer,
-                                                   _nondim_resist_spitzer,
-                                                   _nondim_tec_spitzer,
-                                                   _nondim_tc_e_braginskii,
-                                                   _nondim_tc_i_braginskii,
-                                                   _nondim_visc_e_braginskii,
-                                                   _nondim_visc_i_braginskii,
-                                                   _nondim_resist_braginskii,
-                                                   _nondim_tec_braginskii,
-                                                   _nondim_tc_e_ji_held,
-                                                   _nondim_resist_ji_held,
-                                                   _nondim_tec_ji_held,
-                                                   _nondim_visc_e_ji_held,
-                                                   _nondim_tc_i_ji_held,
-                                                   _nondim_visc_i_ji_held,
-                                                   resistivity,
-                                                   electron_thermal_conductivity,
-                                                   ion_thermal_conductivity,
-                                                   electron_viscosity,
-                                                   ion_viscosity,
-                                                   thermoelectric_conductivity,
+from plasmapy.transport.braginskii import (_nondim_thermal_conductivity,
+                                           _nondim_viscosity,
+                                           _nondim_resistivity,
+                                           _nondim_te_conductivity,
+                                           _check_Z,
+                                           _nondim_tc_e_spitzer,
+                                           _nondim_resist_spitzer,
+                                           _nondim_tec_spitzer,
+                                           _nondim_tc_e_braginskii,
+                                           _nondim_tc_i_braginskii,
+                                           _nondim_visc_e_braginskii,
+                                           _nondim_visc_i_braginskii,
+                                           _nondim_resist_braginskii,
+                                           _nondim_tec_braginskii,
+                                           _nondim_tc_e_ji_held,
+                                           _nondim_resist_ji_held,
+                                           _nondim_tec_ji_held,
+                                           _nondim_visc_e_ji_held,
+                                           _nondim_tc_i_ji_held,
+                                           _nondim_visc_i_ji_held,
+                                           resistivity,
+                                           electron_thermal_conductivity,
+                                           ion_thermal_conductivity,
+                                           electron_viscosity,
+                                           ion_viscosity,
+                                           thermoelectric_conductivity,
                                                    )
-from plasmapy.physics.transport import ClassicalTransport
+from plasmapy.transport import ClassicalTransport
 
 
 def count_decimal_places(digits):
